pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vaptor"
version = "0.1.1"
description = "CLI-based automated VAPT pipeline with Nmap, testssl, Nessus, correlation and reporting"
authors = [
    {name = "Wolfram178"}
]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "requests>=2.31",
    "pandas>=2.0",
    "openpyxl>=3.1",
    "colorama>=0.4.6",
    "cryptography>=42.0",
    "pillow>=10.0"
]

[project.scripts]
vaptor = "vaptor.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = [
    "vaptor*",
    "core*",
    "db*",
    "scanners*",
    "parsers*",
    "output*",
    "utils*",
    "config*",
    "correlator*"
]
